requirements2.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
astor=0.8.0=py37_0
backcall=0.1.0=py37_0
blas=1.0=mkl
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
certifi=2020.4.5.1=py37_0
cffi=1.14.0=py37h2e261b9_0
cudatoolkit=9.2=0
cudnn=7.6.4=cuda9.2_0
cupti=9.2.148=0
cycler=0.10.0=py37_0
dbus=1.13.2=h714fa37_1
decorator=4.4.1=py_0
deprecated=1.2.6=py_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gast=0.3.2=py_0
glib=2.56.2=hd408876_0
grpcio=1.14.1=py37h9ba97e2_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.8.0=py37h989c5e5_3
hdf5=1.10.2=hba1933b_1
icu=58.2=h9c2bf20_1
intel-openmp=2019.0=118
ipython=7.12.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
jedi=0.16.0=py37_0
jpeg=9b=h024ee3a_2
jsonpickle=1.3=py_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.7.1=hd408876_0
libstdcxx-ng=8.2.0=hdf63c60_1
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
markdown=3.1.1=py37_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mock=4.0.1=py_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.4.5=py37_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py_0
openssl=1.0.2u=h7b6447c_0
pandas=1.0.1=py37h0573a6f_0
parso=0.6.1=py_0
pbr=5.1.3=py_0
pcre=8.42=h439df22_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
prompt_toolkit=3.0.3=py_0
protobuf=3.7.1=py37he6710b0_0
psutil=5.6.7=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
pycparser=2.19=py_0
pygments=2.5.2=py_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.0=h6e4f718_3
python-dateutil=2.8.1=py_0
pytorch=1.3.1=cuda92py37hb0ba70e_0
pytz=2019.3=py_0
qt=5.9.6=h8703b6f_2
readline=7.0=h7b6447c_5
scipy=1.4.1=py37h0b6359f_0
setuptools=45.2.0=py37_0
sip=4.19.8=py37hf484d3e_0
six=1.14.0=py37_0
sqlite=3.30.1=h7b6447c_0
tensorboard=1.13.1=py37hf484d3e_0
tensorflow=1.13.1=gpu_py37h49933a4_0
tensorflow-base=1.13.1=gpu_py37h611c6d2_0
tensorflow-estimator=1.13.0=py_0
termcolor=1.1.0=py37_1
tk=8.6.8=hbc83047_0
torchvision=0.4.2=cuda92py37h1667eeb_0
tornado=6.0.3=py37h7b6447c_3
traitlets=4.3.3=py37_0
wcwidth=0.1.8=py_0
werkzeug=0.16.0=py_0
wheel=0.34.2=py37_0
wrapt=1.11.2=py37h7b6447c_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
